* tracker/4.9/camera-android: (59 commits)
  intel-ipu4: fix IPU4 driver crash issue
  intel-ipu4: Added locks around fw closing.
  Media:crlmodule:crl_adv7481_hdmi_configuration.h
  isys & psys package lib2600b0 for commit-id 1d11de3
  isys & psys package lib2600b0 for commit-id b16abb9
  intel-ipu: psys: allocate resource from one pool
  intel-ipu: psys: a fix for NULL pointer
  isys & psys package lib2600b0 for commit-id acca07a
  isys & psys package lib2600b0 for commit-id 9f30e7d
  isys & psys package lib2600b0 for commit-id 36d9833
  isys & psys package lib2600b0 for commit-id 44bfdad
  isys & psys package lib2600b0 for commit-id d05f540
  isys & psys package lib2600b0 for commit-id 03f887e
  isys & psys package lib2600b0 for commit-id c22ec73
  isys & psys package lib2600b0 for commit-id 0299bcc
  isys & psys package lib2600b0 for commit-id 5580304
  isys & psys package lib2600b0 for commit-id a78d9dd
  isys & psys package lib2600b0 for commit-id f986161
  isys & psys package lib2600b0 for commit-id 2d8315a
  intel-ipu4-pdata: imx477: fix GPIOs for LeafHill
  ...

+++ drivers/media/pci/intel-ipu4/intel-ipu4.h
@@ -258,7 +258,7 @@ static inline unsigned int ipu_readl(const volatile void __iomem *addr)
 {
 	unsigned int rval;
 
-	switch ((unsigned int)addr) {
+	switch ((unsigned long)addr) {
 	/* SYSCOM_STATE_READY */
 	case 0x6288008:
 		rval = 0x57A7E001;
